requirements.txt
django-extensions==2.2.6
django-restframework==0.0.1
six==1.14.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Django==3.0.2
django-extensions==2.2.6
django-restframework==0.0.1
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
